+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [3.1.3] - 2025-06-13
+
+### Added
+
+- Added new endpoint: `POST /v3/signature/certificate/{document-number}`.
+
+### Removed
+
+- Removed notification-based certificate choice request with document number.
+
 ## [3.1.2] - 2025-06-05
 
 ### Changed